Fix model

+++ src/model/repeated_campaign/repeated_campaign.go
@@ -8,6 +8,6 @@ type RepeatedCampaign struct {
 	ID        uint  `json:"id"`
 	Start     int64 `json:"start"`
 	End       int64 `json:"end"`
-	Frequency int64 `json:"frequency"`
+	Frequency uint `json:"frequency"`
 	campaign.Campaign
 }
